dvc run of multiple commands
I didn't find any trace of this piece of information in the current documentation. I know that in most cases using `dvc run` twice is the best option, because that way every step has its own versioning, but still I feel this piece of information is currently missing.
I tried adding it to the examples, but I'm really not sure if a CAVEAT should be added here, stating that in most cases running it twice is the best option, or explaining in greater detail the proposed use-case. Maybe the best option is using a more general use case (e.g. "python cmd1.py && cmd2.py") leaving to the reader the research of meaningful use-cases. Really really not sure